backend/uv_requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tinydb==4.8.0
python-dotenv==1.0.1
numpy==1.26.4
google-generativeai==0.7.2
torch==2.3.0+cu121
appdirs==1.4.4